requirements.txt
gradio==4.29.0
ultralytics==8.3.123
opencv-python==4.11.0.86
numpy==2.2.3
Pillow==10.3.0
torch==2.7.0
torchvision==0.22.0
matplotlib==3.10.1
pandas==2.2.3
seaborn==0.13.2
scikit-learn==1.6.1
tqdm==4.67.1
joblib==1.4.2

